requirements.txt
google-generativeai>=0.8.0
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=1.0.0
tavily-python>=0.3.0
edgartools>=2.0.0
pymupdf>=1.24.0
langchain-community>=0.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio-throttle>=1.0.0
rich>=13.0.0
typer>=0.12.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
reportlab>=4.1.0
streamlit>=1.35.0
